pyproject.toml
[project]
name = "DBDPlanner"
version = "2.1.0"
requires-python = ">=3.12"
dependencies = [
    "pip==24.1.2",
    "setuptools==71.0.4",
    "wheel==0.43.0",

    "annotated-types==0.7.0",
    "click==8.1.7",
    "Pillow==10.4.0",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "typing-extensions==4.12.2",
]

[project.optional-dependencies]
installer = ["uv==0.2.27"]
tests = [
    "coverage==7.6.0",
    "iniconfig==2.0.0",
    "packaging==24.1",
    "pluggy==1.5.0",
    "pytest==8.3.1",
    "pytest-cov==5.0.0",
]
dev_tools = [
    "cfgv==3.4.0",
    "distlib==0.3.8",
    "filelock==3.15.4",
    "identify==2.6.0",
    "mypy==1.11.0",
    "mypy-extensions==1.0.0",
    "nodeenv==1.9.1",
    "platformdirs==4.2.2",
    "pre-commit==3.7.1",
    "pyyaml==6.0.1",
    "ruff==0.5.4",
    "virtualenv==20.26.3",
]

[tool.mypy]
python_version = "3.12"
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
exclude = ".venv"

[tool.ruff]
line-length = 79
extend-exclude = [".venv*"]
fix = true
show-fixes = true

[tool.ruff.lint]
select = [
    "A",
    # "AIR",  # there is no Airflow
    "ANN",
    "ARG",
    # "ASYNC",  # may come in handy in future
    "B",
    "BLE",
    "C4",
    "C901",
    "COM",
    "D",
    # "DTZ",  # not sure that needed
    "FBT",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FA",
    # "FIX",  # sometimes it's hardto fix it
    "FLY",
    # "FURB",  # unstable, but come in handy
    "G",
    "I",
    "ICN",
    "INP",
    "INT",
    "ISC",
    "LOG",
    "N",
    # "NPY",  # there is no NumPy
    # "PD",  # there is no Pandas
    "PERF",
    "PGH",
    "PIE",
    "PL",
    # "PT",  # someday I will add some tests...
    "PYI",
    "Q",
    "R",
    "RET",
    "RSE",
    "RUF",
    "PTH",
    "S",  # not sure that needed, but no errors from that
    "SLF",
    "SLOT",
    "SIM",
    "T10",
    "T20",
    "TCH",
    "TD",
    "TID",
    # "TRIO",  # may come in handy in future
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "D100",  # this rule forces to write many docs that I think not required
    "D203",  # just alternative for D211
    "D213",  # just alternative for D212
    "S201",  # it's not Flask app
    "S308",  # it's not Django app
    "S610",  # it's not Django app
    "S611",  # it's not Django app
    "TD003",  # not planning write to do with issues link
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"src/tests/auto/**.py" = ["S101"]  # pytest using asserts

[tool.coverage.run]
branch = true
omit = [
    "src/tests/*",
    "pre-commit.py",
    "create_plan.py",
    "test.py",
]

[tool.coverage.report]
show_missing = true

[tool.coverage.json]
pretty_print = true
